# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Restaurant Management System API
  version: 0.1.0
  description: Proposed MVP contract for the Express/TypeScript implementation. Slice 001 validates the document; each operation becomes implementation-authoritative when its delivery slice adds matching server behavior and contract tests.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api
...
paths:
...
  /v1/staff/restaurants:
    get:
      tags: [Configuration]
      operationId: listRestaurants
      summary: List restaurants visible in the authenticated tenant scope.
      security: [{staffCookie: []}]
      responses:
        "200":
          description: Tenant-scoped restaurants.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Restaurant"}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Configuration]
      operationId: createRestaurant
      summary: Create a restaurant inside the authenticated business account.
      security: [{staffCookie: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateRestaurant"}
      responses:
        "201":
          description: Restaurant created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Restaurant"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
  /v1/staff/restaurants/{restaurantId}:
    parameters:
      - $ref: "#/components/parameters/RestaurantId"
    get:
      tags: [Configuration]
      operationId: getRestaurant
      summary: Get a tenant-scoped restaurant.
      security: [{staffCookie: []}]
      responses:
        "200":
          description: Restaurant.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Restaurant"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Configuration]
      operationId: updateRestaurant
      summary: Edit, activate, or deactivate a restaurant without deleting history.
      security: [{staffCookie: [], csrfHeader: []}]
      parameters:
        - $ref: "#/components/parameters/ExpectedVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateRestaurant"}
      responses:
        "200":
          description: Updated restaurant.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Restaurant"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
  /v1/staff/branches:
    get:
      tags: [Configuration]
      operationId: listAssignedBranches
      summary: List branches assigned to the authenticated employee.
      security: [{staffCookie: []}]
      responses:
        "200":
          description: Assigned branches.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: {type: array, items: {$ref: "#/components/schemas/Branch"}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Configuration]
      operationId: createBranch
      summary: Create and assign a branch in an authorized restaurant.
      security: [{staffCookie: [], csrfHeader: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateBranch"}
      responses:
        "201":
          description: Branch created.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Branch"}
        "404": {$ref: "#/components/responses/NotFound"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
  /v1/staff/branches/{branchId}:
    parameters:
      - $ref: "#/components/parameters/BranchId"
    get:
      tags: [Configuration]
      operationId: getAssignedBranch
      summary: Get an assigned branch without exposing another tenant or branch.
      security: [{staffCookie: []}]
      responses:
        "200":
          description: Branch.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Branch"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Configuration]
      operationId: updateBranch
      summary: Configure branch profile, hours, currency, status, and service availability.
      security: [{staffCookie: [], csrfHeader: []}]
      parameters:
        - $ref: "#/components/parameters/ExpectedVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateBranch"}
      responses:
        "200":
          description: Updated branch.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Branch"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationFailed"}
...
components:
  securitySchemes:
    bootstrapSecret:
      type: apiKey
      in: header
      name: X-Bootstrap-Secret
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    staffCookie:
      type: apiKey
      in: cookie
      name: rms_staff_session
    guestCookie:
      type: apiKey
      in: cookie
      name: rms_guest_session
  parameters:
    RestaurantId:
      name: restaurantId
      in: path
      required: true
      schema: {type: string, format: uuid}
    BranchId:
      name: branchId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    ExpectedVersion:
      name: If-Match
      in: header
      required: true
      description: Quoted integer aggregate version.
      schema: {type: string, pattern: '^"[0-9]+"$'}
...
  responses:
    Unauthorized:
      description: Authentication required.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    Forbidden:
      description: Authenticated but not authorized.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    NotFound:
      description: Resource absent or not visible in the current scope.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    Conflict:
      description: State, version, menu, table, payment, or idempotency conflict.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
    ValidationFailed:
      description: Request validation failed.
      content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}
...
  schemas:
...
    Address:
      type: object
      required: [line1, city, countryCode]
      properties:
        line1: {type: string, minLength: 1, maxLength: 200}
        line2: {type: string, minLength: 1, maxLength: 200}
        city: {type: string, minLength: 1, maxLength: 120}
        region: {type: string, minLength: 1, maxLength: 120}
        postalCode: {type: string, minLength: 1, maxLength: 32}
        countryCode: {type: string, pattern: "^[A-Z]{2}$"}
    ContactInformation:
      type: object
      minProperties: 1
      properties:
        email: {type: string, format: email, maxLength: 320}
        phone: {type: string, minLength: 5, maxLength: 32}
    OpeningPeriod:
      type: object
      required: [dayOfWeek, opensAt, closesAt]
      properties:
        dayOfWeek: {type: integer, minimum: 0, maximum: 6}
        opensAt: {type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"}
        closesAt: {type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"}
    Restaurant:
      type: object
      required: [id, businessAccountId, name, status, branding, settings, version]
      properties:
        id: {type: string, format: uuid}
        businessAccountId: {type: string, format: uuid}
        name: {type: string}
        status: {type: string, enum: [active, inactive]}
        branding: {type: object, additionalProperties: true}
        settings: {type: object, additionalProperties: true}
        version: {type: integer, minimum: 1}
    CreateRestaurant:
      type: object
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        branding: {type: object, additionalProperties: true}
        settings: {type: object, additionalProperties: true}
    UpdateRestaurant:
      type: object
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        status: {type: string, enum: [active, inactive]}
        branding: {type: object, additionalProperties: true}
        settings: {type: object, additionalProperties: true}
    Branch:
      type: object
      required: [id, businessAccountId, restaurantId, name, address, contact, timeZone, currency, status, serviceStatus, allowOrderOverride, version, openingHours]
      properties:
        id: {type: string, format: uuid}
        businessAccountId: {type: string, format: uuid}
        restaurantId: {type: string, format: uuid}
        name: {type: string}
        address: {$ref: "#/components/schemas/Address"}
        contact: {$ref: "#/components/schemas/ContactInformation"}
        timeZone: {type: string}
        currency: {type: string, pattern: "^[A-Z]{3}$"}
        status: {type: string, enum: [active, inactive]}
        serviceStatus: {type: string, enum: [open, closed, temporarily_unavailable]}
        allowOrderOverride: {type: boolean}
        version: {type: integer, minimum: 1}
        openingHours: {type: array, items: {$ref: "#/components/schemas/OpeningPeriod"}}
    CreateBranch:
      type: object
      required: [restaurantId, name, address, contact, timeZone, currency, openingHours]
      properties:
        restaurantId: {type: string, format: uuid}
        name: {type: string, minLength: 1, maxLength: 160}
        address: {$ref: "#/components/schemas/Address"}
        contact: {$ref: "#/components/schemas/ContactInformation"}
        timeZone: {type: string, minLength: 1, maxLength: 100}
        currency: {type: string, pattern: "^[A-Z]{3}$"}
        openingHours: {type: array, minItems: 1, maxItems: 28, items: {$ref: "#/components/schemas/OpeningPeriod"}}
    UpdateBranch:
      type: object
      minProperties: 1
      properties:
        name: {type: string, minLength: 1, maxLength: 160}
        address: {$ref: "#/components/schemas/Address"}
        contact: {$ref: "#/components/schemas/ContactInformation"}
        timeZone: {type: string, minLength: 1, maxLength: 100}
        currency: {type: string, pattern: "^[A-Z]{3}$"}
        status: {type: string, enum: [active, inactive]}
        serviceStatus: {type: string, enum: [open, closed, temporarily_unavailable]}
        allowOrderOverride: {type: boolean}
        openingHours: {type: array, minItems: 1, maxItems: 28, items: {$ref: "#/components/schemas/OpeningPeriod"}}
...
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        code:
          type: string
          enum: [validation_error, authentication_required, permission_denied, feature_disabled, resource_not_found, invalid_state_transition, concurrency_conflict, menu_changed, dish_unavailable, table_unavailable, payment_conflict, idempotency_conflict, rate_limited, service_unavailable, unexpected_error]
        detail: {type: [string, "null"]}
        correlationId: {type: string}
        currentVersion: {type: integer, minimum: 1}
        errors:
          type: object
          additionalProperties:
            type: array
            items: {type: string}